IRR & IRT

+++ fragments/lcd/lcd.ino
@@ -7,7 +7,7 @@
 
 LiquidCrystal_I2C lcd(0x3F, 2, 1, 0, 4, 5, 6, 7, 3, POSITIVE);  // Set the LCD I2C address
 
-const int LED = 10;
+const byte LED = 10;
 
 void setup()  
 {
